+++ source/core/misc/cTimer.h
@@ -11,15 +11,15 @@ namespace vg
 
         void update() { m_last_update = std::chrono::high_resolution_clock::now(); }
 
-        long long getSecondDelta( const bool _update ) { return getNanoDelta( _update ) / 1'000'000'000; }
-        long long getMilliDelta( const bool _update ) { return getNanoDelta( _update ) / 1'000'000; }
-        long long getMicroDelta( const bool _update ) { return getNanoDelta( _update ) / 1'000; }
-        long long getNanoDelta( const bool _update );
+        double getSecondDelta( const bool _update ) { return getNanoDelta( _update ) / 1'000'000'000; }
+        double getMilliDelta( const bool _update ) { return getNanoDelta( _update ) / 1'000'000; }
+        double getMicroDelta( const bool _update ) { return getNanoDelta( _update ) / 1'000; }
+        double getNanoDelta( const bool _update );
 
-        long long getSecondLife() const { return getNanoLife() / 1'000'000'000; }
-        long long getMilliLife() const { return getNanoLife() / 1'000'000; }
-        long long getMicroLife() const { return getNanoLife() / 1'000; }
-        long long getNanoLife() const;
+        double getSecondLife() const { return getNanoLife() / 1'000'000'000; }
+        double getMilliLife() const { return getNanoLife() / 1'000'000; }
+        double getMicroLife() const { return getNanoLife() / 1'000; }
+        double getNanoLife() const;
 
     protected:
         std::chrono::high_resolution_clock::time_point m_start;
